move column mapping to own file
